refactor: clean up display widgets and add caracter card data model
refactor display widgets
