+++ lib/features/fav_characters/presentation/widgets/fav_character_display.dart
@@ -1,5 +1,5 @@
 import 'package:characters_list_app/core/constants/app_constans.dart';
-import 'package:characters_list_app/core/widgets/character_card_with_favorite_status.dart';
+import 'package:characters_list_app/core/widgets/characters/character_card_with_favorite_status.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -10,6 +10,7 @@ class FavCharacterDisplay extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final double height = MediaQuery.of(context).size.height * 0.2;
     return Padding(
       padding: const EdgeInsets.all(5.0),
       child: BlocBuilder<FavCharactersSortingCubit, FavCharactersSortingState>(
@@ -28,6 +29,7 @@ class FavCharacterDisplay extends StatelessWidget {
                   key: ValueKey(character.name),
                   character: character,
                   pageViewTag: favDisplayTag,
+                  height: height,
                 );
               },
             );
